App: split LoanScreen into components
LoanScreen panel components:

- PanelUpdateBorrowPosition
- PanelUpdateLeveragePosition
- PanelUpdateRate
- PanelClosePosition

App UI Components:

- InfoBox
- WarningBox
- InputTokenBadge
- ValueUpdate
