+++ frontend/app/src/screens/LoanScreen/LoanScreen.tsx
@@ -1,46 +1,18 @@
 "use client";
 
 import type { PositionLoan } from "@/src/types";
-import type { ReactNode } from "react";
 
-import { ConnectWarningBox } from "@/src/comps/ConnectWarningBox/ConnectWarningBox";
-import { Field } from "@/src/comps/Field/Field";
-import { LeverageField, useLeverageField } from "@/src/comps/LeverageField/LeverageField";
 import { Position } from "@/src/comps/Position/Position";
 import { Screen } from "@/src/comps/Screen/Screen";
-import { ETH_MAX_RESERVE, INTEREST_RATE_INCREMENT, INTEREST_RATE_MAX, INTEREST_RATE_MIN } from "@/src/constants";
-import content from "@/src/content";
-import { ACCOUNT_BALANCES, ACCOUNT_POSITIONS, getDebtBeforeRateBucketIndex, INTEREST_CHART } from "@/src/demo-mode";
-import { useInputFieldValue } from "@/src/form-utils";
-import { fmtnum, formatRisk } from "@/src/formatting";
-import { getLiquidationPriceFromLeverage, getLoanDetails } from "@/src/liquity-math";
-import { useAccount } from "@/src/services/Ethereum";
-import { usePrice } from "@/src/services/Prices";
-import { infoTooltipProps, riskLevelToStatusMode } from "@/src/uikit-utils";
+import { ACCOUNT_POSITIONS } from "@/src/demo-mode";
 import { css } from "@/styled-system/css";
-import {
-  Button,
-  Checkbox,
-  Dropdown,
-  HFlex,
-  IconSettings,
-  InfoTooltip,
-  InputField,
-  lerp,
-  norm,
-  Slider,
-  StatusDot,
-  Tabs,
-  TextButton,
-  TokenIcon,
-  TOKENS_BY_SYMBOL,
-  VFlex,
-} from "@liquity2/uikit";
-import * as dn from "dnum";
+import { IconSettings, Tabs, VFlex } from "@liquity2/uikit";
 import { notFound, useRouter, useSearchParams, useSelectedLayoutSegment } from "next/navigation";
-import { useEffect, useMemo, useRef, useState } from "react";
-
-const ARROW_RIGHT = "→";
+import { useMemo, useState } from "react";
+import { PanelClosePosition } from "./PanelClosePosition";
+import { PanelUpdateBorrowPosition } from "./PanelUpdateBorrowPosition";
+import { PanelUpdateLeveragePosition } from "./PanelUpdateLeveragePosition";
+import { PanelUpdateRate } from "./PanelUpdateRate";
 
 const TABS = [
   { label: "Collateral & Debt", id: "colldebt" },
@@ -106,1151 +78,14 @@ export function LoanScreen() {
           />
           {action === "colldebt" && (
             leverageMode
-              ? <UpdateLeveragePositionPanel loan={trove} />
-              : <UpdateBorrowPositionPanel loan={trove} />
+              ? <PanelUpdateLeveragePosition loan={trove} />
+              : <PanelUpdateBorrowPosition loan={trove} />
           )}
-          {action === "rate" && <UpdateRatePanel loan={trove} />}
-          {action === "close" && <ClosePositionPanel loan={trove} />}
+          {action === "rate" && <PanelUpdateRate loan={trove} />}
+          {action === "close" && <PanelClosePosition loan={trove} />}
         </VFlex>
       </VFlex>
     </Screen>
-  );
-}
-
-type RelativeFieldMode = "add" | "remove";
-
-function UpdateBorrowPositionPanel({ loan }: { loan: PositionLoan }) {
-  const router = useRouter();
-  const account = useAccount();
-
-  const collateral = TOKENS_BY_SYMBOL[loan.collateral];
-  const collPrice = usePrice(collateral.symbol);
-  const boldPriceUsd = usePrice("BOLD") ?? dn.from(0, 18);
-
-  // deposit change
-  const [depositMode, setDepositMode] = useState<RelativeFieldMode>("add");
-  const depositChange = useInputFieldValue((value) => dn.format(value));
-
-  // deposit update
-  const newDeposit = depositChange.parsed && (
-    depositMode === "remove"
-      ? dn.sub(loan.deposit, depositChange.parsed)
-      : dn.add(loan.deposit, depositChange.parsed)
-  );
-
-  const collMax = depositMode === "remove" ? loan.deposit : dn.sub(
-    ACCOUNT_BALANCES[collateral.symbol],
-    ETH_MAX_RESERVE,
-  );
-
-  // debt change
-  const [debtMode, setDebtMode] = useState<RelativeFieldMode>("add");
-  const debtChange = useInputFieldValue((value) => dn.format(value));
-  const debtChangeUsd = debtChange.parsed && dn.mul(debtChange.parsed, boldPriceUsd);
-
-  const newDebt = debtChange.parsed && (
-    debtMode === "remove"
-      ? dn.sub(loan.borrowed, debtChange.parsed)
-      : dn.add(loan.borrowed, debtChange.parsed)
-  );
-
-  const boldMax = debtMode === "remove" ? ACCOUNT_BALANCES["BOLD"] : null;
-
-  const loanDetails = getLoanDetails(
-    loan.deposit,
-    loan.borrowed,
-    loan.interestRate,
-    collateral.collateralRatio,
-    collPrice,
-  );
-
-  const newLoanDetails = getLoanDetails(
-    newDeposit,
-    newDebt,
-    loanDetails.interestRate,
-    collateral.collateralRatio,
-    collPrice,
-  );
-
-  const allowSubmit = account.isConnected && (
-    !dn.eq(loanDetails.deposit ?? dn.from(0, 18), newLoanDetails.deposit ?? dn.from(0, 18))
-    || !dn.eq(loanDetails.debt ?? dn.from(0, 18), newLoanDetails.debt ?? dn.from(0, 18))
-  );
-
-  return (
-    <>
-      <VFlex gap={32}>
-        <Field
-          field={
-            <InputField
-              {...depositChange.inputFieldProps}
-              contextual={
-                <InputTokenBadge
-                  background={false}
-                  icon={<TokenIcon symbol={collateral.symbol} />}
-                  label={collateral.name}
-                />
-              }
-              label={{
-                start: depositMode === "remove"
-                  ? "Decrease your collateral"
-                  : "Increase your collateral",
-                end: (
-                  <Tabs
-                    compact
-                    items={[
-                      { label: "Deposit", panelId: "panel-deposit", tabId: "tab-deposit" },
-                      { label: "Withdraw", panelId: "panel-withdraw", tabId: "tab-withdraw" },
-                    ]}
-                    onSelect={(index) => {
-                      setDepositMode(index === 1 ? "remove" : "add");
-                      depositChange.setValue("0");
-                    }}
-                    selected={depositMode === "remove" ? 1 : 0}
-                  />
-                ),
-              }}
-              labelHeight={32}
-              placeholder="0.00"
-              secondary={{
-                start: (depositChange.parsed && collPrice)
-                  ? "$" + fmtnum(dn.mul(depositChange.parsed, collPrice))
-                  : "$0.00",
-                end: (
-                  <TextButton
-                    label={`Max ${fmtnum(collMax, 2)} ${TOKENS_BY_SYMBOL[collateral.symbol].name}`}
-                    onClick={() => {
-                      depositChange.setValue(dn.toString(collMax));
-                    }}
-                  />
-                ),
-              }}
-            />
-          }
-          footer={[[
-            <Field.FooterInfo label="Collateral after" />,
-            loanDetails.deposit && newLoanDetails.deposit && (
-              <Field.FooterInfo
-                label={
-                  <HFlex alignItems="center" gap={8}>
-                    <div>{fmtnum(loanDetails.deposit, 2)}</div>
-                    <div>{ARROW_RIGHT}</div>
-                  </HFlex>
-                }
-                value={
-                  <HFlex alignItems="center" gap={8}>
-                    <div>
-                      {fmtnum(newLoanDetails.deposit, 2)} {TOKENS_BY_SYMBOL[collateral.symbol].name}
-                    </div>
-                    <InfoTooltip heading="Collateral update" />
-                  </HFlex>
-                }
-              />
-            ),
-          ]]}
-        />
-
-        <Field
-          field={
-            <InputField
-              {...debtChange.inputFieldProps}
-              contextual={
-                <InputTokenBadge
-                  background={false}
-                  icon={<TokenIcon symbol="BOLD" />}
-                  label="BOLD"
-                />
-              }
-              label={{
-                start: debtMode === "remove"
-                  ? "Decrease your debt"
-                  : "Increase your debt",
-                end: (
-                  <Tabs
-                    compact
-                    items={[
-                      { label: "Borrow", panelId: "panel-borrow", tabId: "tab-borrow" },
-                      { label: "Repay", panelId: "panel-repay", tabId: "tab-repay" },
-                    ]}
-                    onSelect={(index) => {
-                      setDebtMode(index === 1 ? "remove" : "add");
-                      debtChange.setValue("0");
-                    }}
-                    selected={debtMode === "remove" ? 1 : 0}
-                  />
-                ),
-              }}
-              labelHeight={32}
-              placeholder="0.00"
-              secondary={{
-                start: debtChangeUsd
-                  ? "$" + fmtnum(debtChangeUsd)
-                  : "$0.00",
-                end: (
-                  boldMax && (
-                    <TextButton
-                      label={`Max ${fmtnum(boldMax)} BOLD`}
-                      onClick={() => {
-                        debtChange.setValue(dn.toString(boldMax));
-                      }}
-                    />
-                  )
-                ),
-              }}
-            />
-          }
-          footer={[
-            [
-              <Field.FooterInfo label="Debt after" />,
-              loanDetails.debt && newLoanDetails.debt && (
-                <Field.FooterInfo
-                  label={
-                    <HFlex alignItems="center" gap={8}>
-                      <div>{fmtnum(loanDetails.debt)}</div>
-                      <div>{ARROW_RIGHT}</div>
-                    </HFlex>
-                  }
-                  value={
-                    <HFlex alignItems="center" gap={8}>
-                      <div>{fmtnum(newLoanDetails.debt)} BOLD</div>
-                      <InfoTooltip heading="Debt update" />
-                    </HFlex>
-                  }
-                />
-              ),
-            ],
-          ]}
-        />
-
-        <div
-          className={css({
-            paddingTop: 8,
-            paddingBottom: 32,
-          })}
-        >
-          <InfoBox>
-            <HFlex justifyContent="space-between" gap={16}>
-              <div>Liquidation risk</div>
-              <ValueUpdate
-                before={loanDetails.liquidationRisk && (
-                  <HFlex gap={4} justifyContent="flex-start">
-                    <StatusDot mode={riskLevelToStatusMode(loanDetails.liquidationRisk)} />
-                    {formatRisk(loanDetails.liquidationRisk)}
-                  </HFlex>
-                )}
-                after={newLoanDetails.liquidationRisk && (
-                  <HFlex gap={4} justifyContent="flex-start">
-                    <StatusDot mode={riskLevelToStatusMode(newLoanDetails.liquidationRisk)} />
-                    {formatRisk(newLoanDetails.liquidationRisk)}
-                  </HFlex>
-                )}
-              />
-            </HFlex>
-            <HFlex justifyContent="space-between" gap={16}>
-              <div>
-                <abbr title="Loan-to-value ratio">LTV</abbr>
-              </div>
-              <HFlex
-                gap={8}
-                className={css({
-                  fontVariantNumeric: "tabular-nums",
-                })}
-              >
-                {loanDetails.ltv && (
-                  <div
-                    className={css({
-                      color: "contentAlt",
-                    })}
-                  >
-                    {fmtnum(dn.mul(loanDetails.ltv, 100))}%
-                  </div>
-                )}
-                <div
-                  className={css({
-                    color: "contentAlt",
-                  })}
-                >
-                  {ARROW_RIGHT}
-                </div>
-                {newLoanDetails.ltv && (
-                  <div>
-                    {fmtnum(dn.mul(newLoanDetails.ltv, dn.lt(newLoanDetails.ltv, 0) ? 0 : 100))}%
-                  </div>
-                )}
-              </HFlex>
-            </HFlex>
-            <HFlex justifyContent="space-between" gap={16}>
-              <div>Liquidation price</div>
-              <HFlex
-                gap={8}
-                className={css({
-                  fontVariantNumeric: "tabular-nums",
-                })}
-              >
-                {loanDetails.liquidationPrice && (
-                  <div
-                    className={css({
-                      color: "contentAlt",
-                    })}
-                  >
-                    ${fmtnum(loanDetails.liquidationPrice)}
-                  </div>
-                )}
-                <div
-                  className={css({
-                    color: "contentAlt",
-                  })}
-                >
-                  {ARROW_RIGHT}
-                </div>
-                {newLoanDetails.liquidationPrice && (
-                  <div>
-                    ${fmtnum(newLoanDetails.liquidationPrice)}
-                  </div>
-                )}
-              </HFlex>
-            </HFlex>
-          </InfoBox>
-        </div>
-      </VFlex>
-
-      <div
-        style={{
-          display: "flex",
-          flexDirection: "column",
-          justifyContent: "center",
-          gap: 32,
-          width: "100%",
-        }}
-      >
-        <ConnectWarningBox />
-        <Button
-          disabled={!allowSubmit}
-          label="Update position"
-          mode="primary"
-          size="large"
-          wide
-          onClick={() => {
-            router.push("/transactions/update-loan");
-          }}
-        />
-      </div>
-    </>
-  );
-}
-
-function UpdateLeveragePositionPanel({ loan }: { loan: PositionLoan }) {
-  const router = useRouter();
-  const account = useAccount();
-
-  const collateral = TOKENS_BY_SYMBOL[loan.collateral];
-  const collPrice = usePrice(collateral.symbol);
-
-  // loan details before the update
-  const initialLoanDetails = getLoanDetails(
-    loan.deposit,
-    loan.borrowed,
-    loan.interestRate,
-    collateral.collateralRatio,
-    collPrice,
-  );
-
-  // deposit change
-  const [depositMode, setDepositMode] = useState<RelativeFieldMode>("add");
-  const depositChange = useInputFieldValue((value) => dn.format(value));
-  const [userLeverageFactor, setUserLeverageFactor] = useState(initialLoanDetails.leverageFactor ?? 1);
-
-  const newDepositPreLeverage = depositChange.parsed
-    ? (depositMode === "remove"
-      ? dn.sub(initialLoanDetails.depositPreLeverage ?? dn.from(0, 18), depositChange.parsed)
-      : dn.add(initialLoanDetails.depositPreLeverage ?? dn.from(0, 18), depositChange.parsed))
-    : initialLoanDetails.depositPreLeverage;
-
-  const newDeposit = dn.mul(newDepositPreLeverage ?? dn.from(0, 18), userLeverageFactor);
-
-  const totalPositionValue = dn.mul(newDeposit, collPrice ?? dn.from(0, 18));
-
-  const newDebt = dn.sub(
-    totalPositionValue,
-    dn.mul(newDepositPreLeverage ?? dn.from(0, 18), collPrice ?? dn.from(0, 18)),
-  );
-
-  const newLoanDetails = getLoanDetails(
-    newDeposit,
-    newDebt,
-    initialLoanDetails.interestRate,
-    collateral.collateralRatio,
-    collPrice,
-  );
-
-  const liquidationPrice = getLiquidationPriceFromLeverage(
-    userLeverageFactor,
-    collPrice ?? dn.from(0, 18),
-    collateral.collateralRatio,
-  );
-
-  const newDepositUsd = collPrice && dn.mul(newDeposit, collPrice);
-
-  const ltv = newDeposit && newLoanDetails.debt && newDepositUsd && dn.gt(newDepositUsd, 0)
-    ? dn.div(newLoanDetails.debt, dn.mul(newDeposit, newDepositUsd))
-    : null;
-
-  // leverage factor
-  const leverageField = useLeverageField({
-    collPrice: collPrice ?? dn.from(0, 18),
-    collToken: collateral,
-    depositPreLeverage: newDepositPreLeverage,
-    maxLtvAllowedRatio: 1, // allow up to the max. LTV
-  });
-
-  useEffect(() => {
-    if (leverageField.leverageFactor !== userLeverageFactor) {
-      setUserLeverageFactor(leverageField.leverageFactor);
-    }
-  }, [leverageField.leverageFactor]);
-
-  const initialLeverageFactorSet = useRef(false);
-  useEffect(() => {
-    if (initialLoanDetails.leverageFactor && !initialLeverageFactorSet.current) {
-      leverageField.updateLeverageFactor(initialLoanDetails.leverageFactor);
-      initialLeverageFactorSet.current = true;
-    }
-  }, [leverageField.updateLeverageFactor, initialLoanDetails.leverageFactor]);
-
-  const depositMax = depositMode === "remove"
-    ? initialLoanDetails.depositPreLeverage
-    : dn.sub(ACCOUNT_BALANCES[collateral.symbol], ETH_MAX_RESERVE);
-
-  const [agreeToLiquidationRisk, setAgreeToLiquidationRisk] = useState(false);
-
-  const showAgreeToLiquidationRisk = ltv
-    ? dn.gt(ltv, newLoanDetails.maxLtvAllowed)
-    : false;
-
-  const allowSubmit = account.isConnected && (
-    !showAgreeToLiquidationRisk || agreeToLiquidationRisk
-  ) && (
-    !dn.eq(
-      initialLoanDetails.deposit ?? dn.from(0, 18),
-      newLoanDetails.deposit ?? dn.from(0, 18),
-    ) || (
-      initialLoanDetails.leverageFactor !== newLoanDetails.leverageFactor
-    )
-  );
-
-  return (
-    <>
-      <VFlex gap={32}>
-        <Field
-          field={
-            <InputField
-              {...depositChange.inputFieldProps}
-              contextual={
-                <InputTokenBadge
-                  background={false}
-                  icon={<TokenIcon symbol={collateral.symbol} />}
-                  label={collateral.name}
-                />
-              }
-              label={{
-                start: depositMode === "remove"
-                  ? "Decrease your deposit"
-                  : "Increase your deposit",
-                end: (
-                  <Tabs
-                    compact
-                    items={[
-                      { label: "Deposit", panelId: "panel-deposit", tabId: "tab-deposit" },
-                      { label: "Withdraw", panelId: "panel-withdraw", tabId: "tab-withdraw" },
-                    ]}
-                    onSelect={(index) => {
-                      setDepositMode(index === 1 ? "remove" : "add");
-                      depositChange.setValue("0");
-                    }}
-                    selected={depositMode === "remove" ? 1 : 0}
-                  />
-                ),
-              }}
-              labelHeight={32}
-              placeholder="0.00"
-              secondary={{
-                start: collPrice && (
-                  depositChange.parsed
-                    ? "$" + fmtnum(dn.mul(depositChange.parsed, collPrice))
-                    : "$0.00"
-                ),
-                end: depositMax && (
-                  <TextButton
-                    label={`Max ${fmtnum(depositMax)} ${collateral.name}`}
-                    onClick={() => {
-                      depositChange.setValue(dn.toString(depositMax));
-                    }}
-                  />
-                ),
-              }}
-            />
-          }
-          footer={[[
-            <Field.FooterInfo label="Deposit after" />,
-            initialLoanDetails.depositPreLeverage && newDepositPreLeverage && (
-              <Field.FooterInfo
-                value={
-                  <HFlex alignItems="center" gap={8}>
-                    <ValueUpdate
-                      before={
-                        <div
-                          title={`${fmtnum(initialLoanDetails.depositPreLeverage, "full")} ${collateral.name}`}
-                        >
-                          {fmtnum(initialLoanDetails.depositPreLeverage)}
-                        </div>
-                      }
-                      after={
-                        <HFlex alignItems="center" gap={8}>
-                          <div
-                            title={`${fmtnum(newDepositPreLeverage, "full")} ${collateral.name}`}
-                          >
-                            {fmtnum(newDepositPreLeverage)} {collateral.name}
-                          </div>
-                          <InfoTooltip heading="Collateral update" />
-                        </HFlex>
-                      }
-                      fontSize={14}
-                    />
-                  </HFlex>
-                }
-              />
-            ),
-          ]]}
-        />
-
-        <Field
-          field={<LeverageField {...leverageField} />}
-          footer={[
-            [
-              <Field.FooterInfo label="ETH liquidation price" />,
-              <ValueUpdate
-                fontSize={14}
-                before={initialLoanDetails.liquidationPrice && (
-                  `$${fmtnum(initialLoanDetails.liquidationPrice)}`
-                )}
-                after={liquidationPrice && (
-                  `$${fmtnum(liquidationPrice)}`
-                )}
-              />,
-            ],
-            [
-              <Field.FooterInfo label="ETH exposure" />,
-              <ValueUpdate
-                fontSize={14}
-                before={initialLoanDetails.depositPreLeverage && (
-                  <div
-                    title={`${fmtnum(initialLoanDetails.deposit, "full")} ${collateral.name}`}
-                  >
-                    {fmtnum(initialLoanDetails.deposit)} {collateral.name}
-                  </div>
-                )}
-                after={newDepositPreLeverage && (
-                  <div title={`${fmtnum(newLoanDetails.deposit, "full")} ${collateral.name}`}>
-                    {fmtnum(newLoanDetails.deposit)} {collateral.name}
-                  </div>
-                )}
-              />,
-            ],
-            [
-              <Field.FooterInfo label="Leverage" />,
-              <ValueUpdate
-                fontSize={14}
-                before={<>{fmtnum(initialLoanDetails.leverageFactor, "1z")}x</>}
-                after={newLoanDetails.isUnderwater
-                  ? "N/A"
-                  : <>{fmtnum(newLoanDetails.leverageFactor, "1z")}x</>}
-              />,
-            ],
-            [
-              <Field.FooterInfo label="Implied total debt" />,
-              <ValueUpdate
-                fontSize={14}
-                before={initialLoanDetails.debt && (
-                  `${fmtnum(initialLoanDetails.debt)} BOLD`
-                )}
-                after={newLoanDetails.debt && (
-                  `${fmtnum(newLoanDetails.debt)} BOLD`
-                )}
-              />,
-            ],
-          ]}
-        />
-
-        <VFlex
-          gap={16}
-          className={css({
-            paddingTop: 8,
-            paddingBottom: 32,
-          })}
-        >
-          <InfoBox gap={8}>
-            <HFlex justifyContent="space-between" gap={16}>
-              <div>Liquidation risk</div>
-              <ValueUpdate
-                before={initialLoanDetails.liquidationRisk && (
-                  <HFlex gap={4} justifyContent="flex-start">
-                    <StatusDot mode={riskLevelToStatusMode(initialLoanDetails.liquidationRisk)} />
-                    {formatRisk(initialLoanDetails.liquidationRisk)}
-                  </HFlex>
-                )}
-                after={newLoanDetails.liquidationRisk && (
-                  <HFlex gap={4} justifyContent="flex-start">
-                    <StatusDot mode={riskLevelToStatusMode(newLoanDetails.liquidationRisk)} />
-                    {formatRisk(newLoanDetails.liquidationRisk)}
-                  </HFlex>
-                )}
-              />
-            </HFlex>
-            <HFlex
-              justifyContent="space-between"
-              gap={16}
-              className={css({
-                fontSize: 14,
-              })}
-            >
-              <div
-                className={css({
-                  color: "contentAlt",
-                })}
-              >
-                <abbr title="Loan-to-value ratio">LTV</abbr>
-              </div>
-              <ValueUpdate
-                fontSize={14}
-                before={initialLoanDetails.ltv && `${fmtnum(dn.mul(initialLoanDetails.ltv, 100))}%`}
-                after={
-                  <span
-                    className={css({
-                      "--color-negative": "token(colors.negative)",
-                    })}
-                    style={{
-                      color: newLoanDetails.ltv && dn.gt(newLoanDetails.ltv, newLoanDetails.maxLtvAllowed)
-                        ? "var(--color-negative)"
-                        : "inherit",
-                    }}
-                  >
-                    {newLoanDetails.ltv && `${fmtnum(dn.mul(newLoanDetails.ltv, 100))}%`}
-                  </span>
-                }
-              />
-            </HFlex>
-          </InfoBox>
-
-          {newLoanDetails.isUnderwater
-            ? (
-              <WarningBox>
-                <div>
-                  Your position is currently underwater. You need to add at least{" "}
-                  {fmtnum(newLoanDetails.requiredCollateralToRecover)}
-                  {" "}
-                  {collateral.name} to bring it back above water.
-                </div>
-              </WarningBox>
-            )
-            : showAgreeToLiquidationRisk
-            ? (
-              <WarningBox>
-                <div>
-                  The maximum LTV for the position is{"  "}
-                  {fmtnum(dn.mul(newLoanDetails.maxLtv, 100))}%. Your updated position may be liquidated immediately.
-                </div>
-                <label
-                  className={css({
-                    display: "flex",
-                    alignItems: "center",
-                    gap: 12,
-                    cursor: "pointer",
-                  })}
-                >
-                  <Checkbox
-                    checked={agreeToLiquidationRisk}
-                    onChange={(checked) => {
-                      setAgreeToLiquidationRisk(checked);
-                    }}
-                  />
-                  I understand. Let’s continue.
-                </label>
-              </WarningBox>
-            )
-            : null}
-        </VFlex>
-      </VFlex>
-
-      <div
-        style={{
-          display: "flex",
-          flexDirection: "column",
-          justifyContent: "center",
-          gap: 32,
-          width: "100%",
-        }}
-      >
-        <ConnectWarningBox />
-        <Button
-          disabled={!allowSubmit}
-          label="Update position"
-          mode="primary"
-          size="large"
-          wide
-          onClick={() => {
-            router.push("/transactions/update-loan");
-          }}
-        />
-      </div>
-    </>
-  );
-}
-
-function UpdateRatePanel({ loan }: { loan: PositionLoan }) {
-  const router = useRouter();
-  const account = useAccount();
-
-  const collateral = TOKENS_BY_SYMBOL[loan.collateral];
-  const collPrice = usePrice(collateral.symbol);
-
-  const deposit = useInputFieldValue((value) => `${dn.format(value)} ${collateral.symbol}`, {
-    defaultValue: dn.toString(loan.deposit),
-  });
-  const debt = useInputFieldValue((value) => `${dn.format(value)} BOLD`, {
-    defaultValue: dn.toString(loan.borrowed),
-  });
-  const interestRate = useInputFieldValue((value) => `${dn.format(value)} %`, {
-    defaultValue: dn.toString(dn.mul(loan.interestRate, 100)),
-  });
-
-  const loanDetails = getLoanDetails(
-    loan.deposit,
-    loan.borrowed,
-    dn.div(loan.interestRate, 100),
-    collateral.collateralRatio,
-    collPrice,
-  );
-
-  const newLoanDetails = getLoanDetails(
-    deposit.isEmpty ? null : deposit.parsed,
-    debt.isEmpty ? null : debt.parsed,
-    interestRate.parsed && dn.div(interestRate.parsed, 100),
-    collateral.collateralRatio,
-    collPrice,
-  );
-
-  const boldInterestPerYear = interestRate.parsed
-    && debt.parsed
-    && dn.mul(debt.parsed, dn.div(interestRate.parsed, 100));
-
-  const boldRedeemableInFront = dn.format(
-    getDebtBeforeRateBucketIndex(
-      interestRate.parsed
-        ? Math.round((dn.toNumber(interestRate.parsed) - INTEREST_RATE_MIN) / INTEREST_RATE_INCREMENT)
-        : 0,
-    ),
-    { compact: true },
-  );
-
-  const allowSubmit = account.isConnected
-    && deposit.parsed
-    && dn.gt(deposit.parsed, 0)
-    && debt.parsed
-    && dn.gt(debt.parsed, 0)
-    && interestRate.parsed
-    && dn.gt(interestRate.parsed, 0);
-
-  return (
-    <>
-      <Field
-        // “Interest rate”
-        field={
-          <InputField
-            contextual={
-              <div
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                  width: 300,
-                }}
-              >
-                <Slider
-                  gradient={[1 / 3, 2 / 3]}
-                  chart={INTEREST_CHART}
-                  onChange={(value) => {
-                    interestRate.setValue(
-                      String(Math.round(lerp(INTEREST_RATE_MIN, INTEREST_RATE_MAX, value) * 10) / 10),
-                    );
-                  }}
-                  value={norm(
-                    interestRate.parsed ? dn.toNumber(interestRate.parsed) : 0,
-                    INTEREST_RATE_MIN,
-                    INTEREST_RATE_MAX,
-                  )}
-                />
-              </div>
-            }
-            label={content.borrowScreen.interestRateField.label}
-            placeholder="0.00"
-            secondary={{
-              start: (
-                <HFlex gap={4}>
-                  <div>
-                    {boldInterestPerYear
-                      ? fmtnum(boldInterestPerYear, 2)
-                      : "−"} BOLD / year
-                  </div>
-                  <InfoTooltip {...infoTooltipProps(content.borrowScreen.infoTooltips.interestRateBoldPerYear)} />
-                </HFlex>
-              ),
-              end: (
-                <span>
-                  <span>{"Before you "}</span>
-                  <span
-                    className={css({
-                      color: "content",
-                    })}
-                  >
-                    <span
-                      style={{
-                        fontVariantNumeric: "tabular-nums",
-                      }}
-                    >
-                      {boldRedeemableInFront}
-                    </span>
-                    <span>{" BOLD to redeem"}</span>
-                  </span>
-                </span>
-              ),
-            }}
-            {...interestRate.inputFieldProps}
-            valueUnfocused={(!interestRate.isEmpty && interestRate.parsed)
-              ? (
-                <span
-                  style={{
-                    display: "flex",
-                    alignItems: "center",
-                    gap: 12,
-                  }}
-                >
-                  <span
-                    style={{
-                      fontVariantNumeric: "tabular-nums",
-                    }}
-                  >
-                    {fmtnum(interestRate.parsed, "1z")}
-                  </span>
-                  <span
-                    style={{
-                      color: "#878AA4",
-                      fontSize: 24,
-                    }}
-                  >
-                    % per year
-                  </span>
-                </span>
-              )
-              : null}
-          />
-        }
-      />
-
-      <div
-        className={css({
-          padding: "8px 0",
-        })}
-      >
-        <InfoBox>
-          <HFlex justifyContent="space-between" gap={16}>
-            <div>Redemption risk</div>
-            <ValueUpdate
-              before={loanDetails.redemptionRisk && (
-                <HFlex gap={4} justifyContent="flex-start">
-                  <StatusDot mode={riskLevelToStatusMode(loanDetails.redemptionRisk)} />
-                  {formatRisk(loanDetails.redemptionRisk)}
-                </HFlex>
-              )}
-              after={newLoanDetails.redemptionRisk && (
-                <HFlex gap={4} justifyContent="flex-start">
-                  <StatusDot mode={riskLevelToStatusMode(newLoanDetails.redemptionRisk)} />
-                  {formatRisk(newLoanDetails.redemptionRisk)}
-                </HFlex>
-              )}
-            />
-          </HFlex>
-          <HFlex justifyContent="space-between" gap={16}>
-            <HFlex gap={4}>
-              <div>Interest rate / day</div>
-              <InfoTooltip heading="Interest rate / day" />
-            </HFlex>
-            {boldInterestPerYear && (
-              <HFlex
-                gap={8}
-                className={css({
-                  fontVariantNumeric: "tabular-nums",
-                })}
-              >
-                ~{fmtnum(dn.div(boldInterestPerYear, 365))} BOLD
-              </HFlex>
-            )}
-          </HFlex>
-          <HFlex
-            justifyContent="space-between"
-            gap={16}
-            className={css({
-              marginTop: -12,
-              fontSize: 14,
-              color: "contentAlt",
-            })}
-          >
-            <div>Annual interest rate is charged daily on the debt</div>
-            <div>per day</div>
-          </HFlex>
-        </InfoBox>
-      </div>
-
-      <div
-        style={{
-          display: "flex",
-          flexDirection: "column",
-          justifyContent: "center",
-          gap: 32,
-          width: "100%",
-        }}
-      >
-        <ConnectWarningBox />
-        <Button
-          disabled={!allowSubmit}
-          label="Update position"
-          mode="primary"
-          size="large"
-          wide
-          onClick={() => {
-            router.push("/transactions/update-loan");
-          }}
-        />
-      </div>
-    </>
-  );
-}
-
-function ClosePositionPanel({ loan }: { loan: PositionLoan }) {
-  const router = useRouter();
-  const account = useAccount();
-  const collPrice = usePrice(loan.collateral);
-  const boldPriceUsd = usePrice("BOLD");
-  const [tokenIndex, setTokenIndex] = useState(0);
-
-  const collateral = TOKENS_BY_SYMBOL[loan.collateral];
-
-  if (!collPrice || !boldPriceUsd) {
-    return null;
-  }
-
-  const loanDetails = getLoanDetails(
-    loan.deposit,
-    loan.borrowed,
-    loan.interestRate,
-    collateral.collateralRatio,
-    collPrice,
-  );
-
-  const repayWith = tokenIndex === 0 ? "BOLD" : collateral.symbol;
-
-  const amountToRepay = repayWith === "BOLD"
-    ? (loanDetails.debt ?? dn.from(0))
-    : (dn.div(loanDetails.debt ?? dn.from(0), collPrice));
-
-  const collToReclaim = repayWith === "BOLD"
-    ? loan.deposit
-    : dn.sub(loan.deposit, amountToRepay);
-
-  const allowSubmit = account.isConnected;
-
-  return (
-    <>
-      <VFlex gap={48}>
-        <Field
-          field={
-            <div
-              className={css({
-                display: "flex",
-                alignItems: "center",
-                gap: 16,
-                justifyContent: "space-between",
-              })}
-            >
-              <div
-                className={css({
-                  display: "flex",
-                  gap: 16,
-                  fontSize: 28,
-                  lineHeight: 1,
-                })}
-              >
-                <div>{fmtnum(amountToRepay)}</div>
-              </div>
-              <Dropdown
-                buttonDisplay={() => ({
-                  label: (
-                    <>
-                      {TOKENS_BY_SYMBOL[(["BOLD", collateral.symbol] as const)[tokenIndex]].name}
-                      <span
-                        className={css({
-                          color: "contentAlt",
-                          fontWeight: 400,
-                        })}
-                      >
-                        {TOKENS_BY_SYMBOL[(["BOLD", collateral.symbol] as const)[tokenIndex]].symbol === "BOLD"
-                          ? " account"
-                          : " loan"}
-                      </span>
-                    </>
-                  ),
-                  icon: <TokenIcon symbol={(["BOLD", collateral.symbol] as const)[tokenIndex]} />,
-                })}
-                items={(["BOLD", collateral.symbol] as const).map((symbol) => ({
-                  icon: <TokenIcon symbol={symbol} />,
-                  label: (
-                    <>
-                      {collateral.name} {symbol === "BOLD" ? "(account balance)" : "(loan collateral)"}
-                    </>
-                  ),
-                }))}
-                menuWidth={300}
-                onSelect={setTokenIndex}
-                selected={tokenIndex}
-              />
-            </div>
-          }
-          footer={[[
-            <Field.FooterInfo
-              label={`$${fmtnum(dn.mul(loan.borrowed, boldPriceUsd), 2)}`}
-              value={null}
-            />,
-            null,
-          ]]}
-          label="You repay with"
-        />
-        <Field
-          field={
-            <div
-              className={css({
-                display: "flex",
-                alignItems: "center",
-                gap: 16,
-                justifyContent: "space-between",
-              })}
-            >
-              <div
-                className={css({
-                  display: "flex",
-                  gap: 16,
-                  fontSize: 28,
-                  lineHeight: 1,
-                })}
-              >
-                <div>{fmtnum(collToReclaim)}</div>
-              </div>
-              <div>
-                <div
-                  className={css({
-                    display: "flex",
-                    alignItems: "center",
-                    gap: 8,
-                    height: 40,
-                    padding: "0 16px 0 8px",
-                    fontSize: 24,
-                    background: "fieldSurface",
-                    borderRadius: 20,
-                    userSelect: "none",
-                  })}
-                >
-                  <TokenIcon symbol={collateral.symbol} />
-                  <div>{collateral.name}</div>
-                </div>
-              </div>
-            </div>
-          }
-          label="You reclaim"
-          footer={[[
-            <Field.FooterInfo
-              label={`$${fmtnum(dn.mul(loan.deposit, collPrice), 2)}`}
-              value={null}
-            />,
-            null,
-          ]]}
-        />
-      </VFlex>
-      <div
-        className={css({
-          padding: 20,
-          textAlign: "center",
-          background: "yellow:200",
-          borderRadius: 8,
-        })}
-      >
-        You are repaying your debt and closing the position. {repayWith === "BOLD"
-          ? `The deposit will be returned to your wallet.`
-          : `To close yor position, a part of your collateral will be sold to pay back the debt. The rest of your collateral will be returned to your wallet.`}
-      </div>
-      <div
-        style={{
-          display: "flex",
-          flexDirection: "column",
-          justifyContent: "center",
-          gap: 32,
-          width: "100%",
-        }}
-      >
-        <ConnectWarningBox />
-        <Button
-          disabled={!allowSubmit}
-          label="Repay & close"
-          mode="primary"
-          size="large"
-          wide
-          onClick={() => {
-            router.push("/transactions/close-loan");
-          }}
-        />
-      </div>
-    </>
-  );
-}
-
-function InputTokenBadge({
-  background = true,
-  icon,
-  label,
-}: {
-  background?: boolean;
-  icon?: ReactNode;
-  label: ReactNode;
-}) {
-  return (
-    <div
-      className={css({
-        display: "flex",
-        alignItems: "center",
-        gap: 8,
-        height: 40,
-        fontSize: 24,
-        userSelect: "none",
-      })}
-      style={{
-        paddingLeft: icon ? 8 : 16,
-        background: background ? "#FFF" : "transparent",
-        borderRadius: background ? 20 : 0,
-        padding: background ? "0 16px" : 0,
-      }}
-    >
-      {icon}
-      <div>
-        {label}
-      </div>
-    </div>
   );
 }
 
@@ -1270,80 +105,4 @@ function useTrove(troveId: string | null) {
     ) && position.troveId === troveIdInt)) ?? null;
     return position as PositionLoan | null;
   }, [troveId]);
-}
-
-function ValueUpdate({
-  after,
-  before,
-  fontSize = 16,
-  tabularNums = true,
-}: {
-  after: ReactNode;
-  before: ReactNode;
-  fontSize?: number;
-  tabularNums?: boolean;
-}) {
-  return (
-    <HFlex
-      gap={8}
-      style={{
-        fontSize,
-        fontVariantNumeric: tabularNums ? "tabular-nums" : "inherit",
-      }}
-    >
-      <div className={css({ color: "contentAlt" })}>{before}</div>
-      <div className={css({ color: "contentAlt" })}>{ARROW_RIGHT}</div>
-      <div>{after}</div>
-    </HFlex>
-  );
-}
-
-function InfoBox({
-  children,
-  gap = 16,
-}: {
-  children: ReactNode;
-  gap?: number;
-}) {
-  return (
-    <div
-      className={css({
-        display: "flex",
-        flexDirection: "column",
-        padding: 16,
-        fontSize: 16,
-        background: "infoSurface",
-        border: "1px solid token(colors.infoSurfaceBorder)",
-        borderRadius: 8,
-      })}
-      style={{
-        gap,
-      }}
-    >
-      {children}
-    </div>
-  );
-}
-
-function WarningBox({
-  children,
-}: {
-  children: ReactNode;
-}) {
-  return (
-    <div
-      className={css({
-        display: "flex",
-        flexDirection: "column",
-        gap: 16,
-        padding: 16,
-        fontSize: 16,
-        color: "negativeContent",
-        background: "negativeStrong",
-        borderRadius: 8,
-      })}
-    >
-      {children}
-    </div>
-  );
 }
